Split coauthors bits into coauthors.go
